chore: rename core/tracker.py to tracker/run.py for clarity
