+++ packages/devqubit-braket/src/devqubit_braket/snapshot.py
@@ -35,7 +35,7 @@ from devqubit_engine.utils.time_utils import utc_now_iso
 
 
 if TYPE_CHECKING:
-    from devqubit_engine.core.tracker import Run
+    from devqubit_engine.core.run import Run
 
 logger = logging.getLogger(__name__)
