+++ packages/devqubit-cirq/src/devqubit_cirq/snapshot.py
@@ -19,7 +19,7 @@ from devqubit_engine.utils.time_utils import utc_now_iso
 
 
 if TYPE_CHECKING:
-    from devqubit_engine.core.tracker import Run
+    from devqubit_engine.core.run import Run
 
 logger = logging.getLogger(__name__)
