+++ packages/devqubit-cirq/tests/test_cirq_adapter.py
@@ -15,7 +15,7 @@ from devqubit_cirq.adapter import (
     _compute_circuit_hash,
     _materialize_circuits,
 )
-from devqubit_engine.core.tracker import track
+from devqubit_engine.core.run import track
 
 
 def _artifacts_of_kind(loaded, kind: str):
