+++ packages/devqubit-engine/tests/test_tracker.py
@@ -5,7 +5,7 @@
 
 from __future__ import annotations
 
-from devqubit_engine.core.tracker import track
+from devqubit_engine.core.run import track
 
 
 class TestBasicTracking:
