+++ packages/devqubit-qiskit-runtime/src/devqubit_qiskit_runtime/snapshot.py
@@ -36,7 +36,7 @@ from devqubit_qiskit_runtime.utils import (
 
 
 if TYPE_CHECKING:
-    from devqubit_engine.core.tracker import Run
+    from devqubit_engine.core.run import Run
 
 logger = logging.getLogger(__name__)
