+++ packages/devqubit-qiskit-runtime/tests/test_qiskit_runtime_adapter.py
@@ -12,7 +12,7 @@ from __future__ import annotations
 
 import json
 
-from devqubit_engine.core.tracker import track
+from devqubit_engine.core.run import track
 from devqubit_qiskit_runtime.adapter import QiskitRuntimeAdapter
