+++ src/devqubit/__init__.py
@@ -98,7 +98,7 @@ if TYPE_CHECKING:
     from devqubit_engine.compare.verify import verify, verify_against_baseline
     from devqubit_engine.core.config import Config, get_config, set_config
     from devqubit_engine.core.record import RunRecord
-    from devqubit_engine.core.tracker import Run, track, wrap_backend
+    from devqubit_engine.core.run import Run, track, wrap_backend
     from devqubit_engine.core.types import ArtifactRef
     from devqubit_engine.storage.factory import create_registry, create_store
     from devqubit_ui.app import run_server
